feat: use default registry and fix sending queries with fragments
