+++ packages/graphiql-plugin-batch-request/index.html
@@ -15,7 +15,6 @@
     </style>
     <link rel="stylesheet" href="https://unpkg.com/graphiql/graphiql.min.css" />
   </head>
-
   <body>
     <div id="root"></div>
     <script type="module" src="../src/test-plugin.tsx"></script>
